requirements.txt
Flask~=3.0.3
pyshark~=0.6
Flask-Cors
docker~=7.0.0
Flask-SocketIO
eventlet
watchdog~=4.0.0
Werkzeug~=3.0.2
pymongo~=4.7.0
python-bsonjs
paramiko~=3.4.0
scp~=0.14.5
celery~=5.4.0